Update dtype and modify DAG model operations
- Change dtype from "float32" to "bfloat16" in train configuration.
- Comment out unused log-space operation functions in `dag_model.py`.
- Update operation names to only include "identity".

+++ config/train_daggpt.py
@@ -35,8 +35,7 @@ min_lr = 3e-5
 dataset = "proofpile"
 
 backend = "nccl"
-# dtype = "bfloat16"
-dtype = "float32"
+dtype = "bfloat16"
 compile = False  # Currently broken for dag model
 
 always_save_checkpoint = True
